+++ pkg/models/invoices/nota_credito_debito_descuento_test.go
@@ -246,7 +246,7 @@ func TestNotaCreditoDebitoDescuento_ElectronicaAll(t *testing.T) {
 func emitirNotaDescuentoIndividual(t *testing.T, tc *TestContext, cuis, cufd, cufdControl string, nroNota int) {
 	fecha := time.Now()
 	// Sector 47: Nota Crédito Débito Descuento, Tipo Doc 3
-	cuf, err := utils.GenerarCUF(tc.Nit, fecha, tc.Sucursal, tc.Modalidad, siat.EmisionOnline, 3, 47, nroNota, tc.PuntoVenta, cufdControl)
+	cuf, err := utils.GenerarCUF(tc.Nit, fecha, tc.Sucursal, tc.Modalidad, siat.EmisionOnline, 3, 47, int64(nroNota), tc.PuntoVenta, cufdControl)
 	if err != nil {
 		t.Fatalf("error al generar CUF: %v", err)
 	}
@@ -446,7 +446,7 @@ func TestNotaCreditoDebitoDescuento_ComputarizadaAll(t *testing.T) {
 func emitirNotaDescuentoComputarizadaIndividual(t *testing.T, tc *TestContext, cuis, cufd, cufdControl string, nroNota int) {
 	fecha := time.Now()
 	// Sector 47: Nota Crédito Débito Descuento Computarizada
-	cuf, err := utils.GenerarCUF(tc.Nit, fecha, tc.Sucursal, tc.Modalidad, siat.EmisionOnline, 3, 47, nroNota, tc.PuntoVenta, cufdControl)
+	cuf, err := utils.GenerarCUF(tc.Nit, fecha, tc.Sucursal, tc.Modalidad, siat.EmisionOnline, 3, 47, int64(nroNota), tc.PuntoVenta, cufdControl)
 	if err != nil {
 		t.Fatalf("error al generar CUF: %v", err)
 	}
